separate log parsers from toysolver-check to ToySolver.SAT.LogParser
