+++ app/toysolver_check.hs
@@ -27,6 +27,7 @@ import Text.Printf
 import qualified ToySolver.FileFormat as FF
 import qualified ToySolver.FileFormat.CNF as CNF
 import qualified ToySolver.SAT.Types as SAT
+import ToySolver.SAT.LogParser (parseSATLog, parseMaxSATLog, parsePBLog)
 import ToySolver.Internal.Util (setEncodingChar8)
 import ToySolver.Version
 
@@ -324,49 +325,6 @@ showMIPExpr e = intercalate " "
   [ intercalate "*" (((if c >= 0 then "+" ++ show c else show c) : map (T.unpack . MIP.varName) vs))
   | MIP.Term c vs <- MIP.terms e
   ]
-
-parsePBLog :: String -> SAT.Model
-parsePBLog s = array (1, maximum (0 : map fst ls2)) ls2
-  where
-    ls = lines s
-    ls2 = do
-      l <- ls
-      case l of
-        'v':xs -> do
-          w <- words xs
-          case w of
-            '-':'x':ys -> return (read ys, False)
-            'x':ys -> return (read ys, True)
-        _ -> mzero
-
-parseMaxSATLog :: String -> SAT.Model
-parseMaxSATLog s =
-  case f (lines s) Nothing [] of
-    (obj, vlines) ->
-      let tmp = [c | c <- concat vlines, not (isSpace c)]
-       in if all (\c -> c == '0' || c == '1') tmp then
-            array (1, length tmp) [(v, c=='1') | (v, c) <- zip [1..] tmp]
-          else
-            let ys = [if l > 0 then (l, True) else (-l, False) | vline <- vlines, w <- words vline, let l = read w]
-             in array (1, maximum (0 : map fst ys)) ys
-  where
-    f :: [String] -> Maybe Integer -> [String] -> (Maybe Integer, [String])
-    f [] obj vlines = (obj, reverse vlines)
-    f (l : ls) obj vlines =
-      case l of
-        'o':xs -> f ls (Just (read (dropWhile isSpace xs))) []
-        'v':xs -> f ls obj (dropWhile isSpace xs : vlines)
-        _ -> f ls obj vlines
-
-parseSATLog :: String -> SAT.Model
-parseSATLog s = array (1, maximum (0 : map fst ys)) ys
-  where
-    xs = do
-      l <- lines s
-      case l of
-        'v':xs -> map read (words xs)
-        _ -> mzero
-    ys = [if y > 0 then (y, True) else (-y, False) | y <- takeWhile (0 /=) xs]
 
 isIntegral :: RealFrac r => MIP.Tol r -> r -> Bool
 isIntegral tol x = abs (x - fromIntegral (floor (x + 0.5) :: Integer)) <= MIP.integralityTol tol
